refactor: modalContext 로직 분리
