+++ src/components/organisms/modal/InputModal.tsx
@@ -2,13 +2,10 @@ import dynamic from "next/dynamic";
 import Image from "next/image";
 import {
   ButtonHTMLAttributes,
-  createContext,
   InputHTMLAttributes,
   LabelHTMLAttributes,
   ReactNode,
-  useContext,
   useRef,
-  useState,
 } from "react";
 import ReactDOM from "react-dom";
 import { useFormContext } from "react-hook-form";
@@ -16,36 +13,10 @@ import { useFormContext } from "react-hook-form";
 import Button from "@/components/atoms/button/Button";
 import Input from "@/components/atoms/input/Input";
 import PlusIcon from "@/components/atoms/plus-icon/PlusIcon";
+import { useModalContext } from "@/contexts/InputModalContext";
 import useDragAndDrop from "@/hooks/useDragAndDrop";
 import useFilePreview from "@/hooks/useFilePreview";
 import { cn } from "@/utils/cn";
-
-interface ModalContextType {
-  isOpen: boolean;
-  openModal: () => void;
-  closeModal: () => void;
-}
-
-const InputModalContext = createContext<ModalContextType | null>(null);
-
-export function InputModalProvider({ children }: { children: ReactNode }) {
-  const [isOpen, setIsOpen] = useState(false);
-
-  const openModal = () => setIsOpen(true);
-  const closeModal = () => setIsOpen(false);
-
-  return (
-    <InputModalContext.Provider value={{ isOpen, openModal, closeModal }}>
-      {children}
-    </InputModalContext.Provider>
-  );
-}
-
-export function useModalContext() {
-  const context = useContext(InputModalContext);
-  if (!context) throw new Error("InputModalProvider 내부에서 사용해야 합니다.");
-  return context;
-}
 
 export default function InputModal({ children }: { children: ReactNode }) {
   const { isOpen } = useModalContext();
